fix(release): harden v0.8.0 validation and artifact checks
- Move exact revised-simplex intersection solving behind a private geometry
  utility while preserving fallback and certificate behavior.
- Harden Level 4 shared-facet and crossing classification through 5D,
  including singular and high-range barycentric systems.
- Compare paper artifacts structurally across platforms and align Tectonic
  prerequisites without pkgx.
- Refresh validation artifacts and uv notebook dependencies, and enforce
  native Python 3.14 annotation semantics.

Closes #506
