+++ tests/semgrep/scripts/tests/python_style.py
@@ -1,0 +1,7 @@
+"""Fixtures for repository Python style rules."""
+
+# ruleid: delaunay.python.no-future-annotations-on-python314
+from __future__ import annotations
+
+# ok: delaunay.python.no-future-annotations-on-python314
+from typing import TYPE_CHECKING
